profile page moves to user controller
