+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users
   root to: "pages#home"
-  get "profile", to: "pages#profile"
+  resources :users, only: %i[show]
 
   get "popular", to: "events#popular"
